rfp_analyzer/pyproject.toml
[project]
name = "rfp-analyzer"
version = "1.0.0"
description = "Intelligent RFP Analysis System with RAG + LLM"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langchain-chroma>=0.1.0",
    "langchain-text-splitters>=0.3.0",
    "langchain-experimental>=0.3.0",
    "langgraph>=0.2.0",
    "langsmith>=0.1.0",
    "openai>=1.50.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "chromadb>=0.5.0",
    "pypdf>=5.0.0",
    "pymupdf>=1.24.0",
    "tiktoken>=0.7.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "httpx>=0.27.0",
    "structlog>=24.4.0",
    "tenacity>=9.0.0",
    "numpy>=1.26.0",
    "python-multipart>=0.0.9",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.8.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "pre-commit>=3.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

